refactor: Encapsulated device detection logic into hook
